refactor(items): move to outer scope
