+++ app/components/items/image.skeleton.tsx
