+++ app/components/items/image.tsx
