+++ app/components/items/top-card.tsx
