+++ app/profile/[user-id]/@history/loading.tsx
@@ -1,4 +1,4 @@
-import { ItemsListSkeleton } from '@app/profile/components/items'
+import { ItemsListSkeleton } from '@app/components/items'
 import { DefaultSection } from '@app/sections'
 import { View } from '@app/types'
